+++ src/RestController.php
@@ -368,6 +368,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $value = $this->getResource()->create($data);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -400,6 +402,7 @@ class RestController extends AbstractRestfulController
             $response = $this->getResponse();
             $response->setStatusCode(201);
             $response->getHeaders()->addHeaderLine('Location', $url);
+            $response->getHeaders()->addHeaderLine('Content-Location', $url);
         }
 
         $events->trigger('create.post', $this, [
@@ -424,6 +427,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $result = $this->getResource()->delete($id);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -455,6 +460,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $result = $this->getResource()->deleteList($data);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -487,6 +494,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $entity = $this->getResource()->fetch($id);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -520,6 +529,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $collection = $this->getResource()->fetchAll();
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -631,6 +642,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $entity = $this->getResource()->patch($id, $data);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -666,6 +679,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $entity = $this->getResource()->update($id, $data);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -700,6 +715,8 @@ class RestController extends AbstractRestfulController
 
         try {
             $collection = $this->getResource()->patchList($data);
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -733,6 +750,8 @@ class RestController extends AbstractRestfulController
             $collection = $this->getResource()->replaceList($data);
         } catch (Exception\InvalidArgumentException $e) {
             return new ApiProblem(400, $e->getMessage());
+        } catch (\Throwable $e) {
+            return $this->createApiProblemFromException($e);
         } catch (\Exception $e) {
             return $this->createApiProblemFromException($e);
         }
@@ -948,9 +967,9 @@ class RestController extends AbstractRestfulController
         $collection->setRouteIdentifierName($this->getRouteIdentifierName());
         $collection->setEntityRoute($this->route);
         $collection->setCollectionName($this->collectionName);
-        $collection->setPageSize($this->getPageSize());
 
         try {
+            $collection->setPageSize($this->getPageSize());
             $collection->setPage($this->getRequest()->getQuery('page', 1));
         } catch (HalInvalidArgumentException $e) {
             return new ApiProblem(400, $e->getMessage());
@@ -965,8 +984,8 @@ class RestController extends AbstractRestfulController
      */
     protected function createHalEntity($entity)
     {
-        if ($entity instanceof HalEntity &&
-            ($entity->getLinks()->has('self') || ! $entity->id)
+        if ($entity instanceof HalEntity
+            && ($entity->getLinks()->has('self') || ! $entity->getId())
         ) {
             return $entity;
         }