+++ test/CollectionIntegrationTest.php
@@ -6,17 +6,21 @@
 
 namespace ZFTest\Rest;
 
+use Interop\Container\ContainerInterface;
 use PHPUnit_Framework_TestCase as TestCase;
-use ReflectionMethod;
+use Zend\EventManager\EventManager;
+use Zend\EventManager\SharedEventManager;
 use Zend\Http\PhpEnvironment\Request;
 use Zend\Http\PhpEnvironment\Response;
 use Zend\Mvc\Controller\ControllerManager;
 use Zend\Mvc\Controller\PluginManager as ControllerPluginManager;
 use Zend\Mvc\MvcEvent;
-use Zend\Mvc\Router\Http\TreeRouteStack;
-use Zend\Mvc\Router\RouteMatch;
+use Zend\Mvc\Router\RouteMatch as V2RouteMatch;
+use Zend\Mvc\Service\EventManagerFactory;
 use Zend\Paginator\Adapter\ArrayAdapter as ArrayPaginator;
 use Zend\Paginator\Paginator;
+use Zend\Router\RouteMatch;
+use Zend\ServiceManager\Factory\InvokableFactory;
 use Zend\ServiceManager\ServiceManager;
 use Zend\Stdlib\Parameters;
 use Zend\Uri;
@@ -27,6 +31,7 @@ use ZF\ApiProblem\View\ApiProblemRenderer;
 use ZF\ContentNegotiation\AcceptListener;
 use ZF\Hal\Extractor\LinkCollectionExtractor;
 use ZF\Hal\Extractor\LinkExtractor;
+use ZF\Hal\Link\LinkUrlBuilder;
 use ZF\Hal\Plugin\Hal as HalHelper;
 use ZF\Hal\View\HalJsonRenderer;
 use ZF\Rest\Resource;
@@ -37,6 +42,8 @@ use ZF\Rest\RestController;
  */
 class CollectionIntegrationTest extends TestCase
 {
+    use TreeRouteStackFactoryTrait;
+
     public function setUp()
     {
         $this->setUpRenderer();
@@ -57,18 +64,24 @@ class CollectionIntegrationTest extends TestCase
         $serverUrlHelper->setScheme('http');
         $serverUrlHelper->setHost('localhost.localdomain');
 
-        $this->linksHelper = $linksHelper = new HalHelper();
-        $linksHelper->setUrlHelper($urlHelper);
-        $linksHelper->setServerUrlHelper($serverUrlHelper);
+        $linkUrlBuilder = new LinkUrlBuilder($serverUrlHelper, $urlHelper);
 
-        $linkExtractor = new LinkExtractor($serverUrlHelper, $urlHelper);
+        $this->linksHelper = $linksHelper = new HalHelper();
+        $linksHelper->setLinkUrlBuilder($linkUrlBuilder);
+
+        $linkExtractor = new LinkExtractor($linkUrlBuilder);
         $linkCollectionExtractor = new LinkCollectionExtractor($linkExtractor);
         $linksHelper->setLinkCollectionExtractor($linkCollectionExtractor);
 
-        $this->helpers = $helpers = new HelperPluginManager();
+        $this->helpers = $helpers = new HelperPluginManager(
+            $this->prophesize(ContainerInterface::class)->reveal()
+        );
         $helpers->setService('url', $urlHelper);
         $helpers->setService('serverUrl', $serverUrlHelper);
         $helpers->setService('hal', $linksHelper);
+        if (method_exists($helpers, 'configure')) {
+            $helpers->setAlias('Hal', 'hal');
+        }
     }
 
     public function setUpRenderer()
@@ -97,11 +110,11 @@ class CollectionIntegrationTest extends TestCase
                 ],
             ],
         ];
-        $this->router = $router = new TreeRouteStack();
+        $this->router = $router = $this->createTreeRouteStack();
         $router->addRoutes($routes);
 
         $matches = $router->match($this->request);
-        if (!$matches instanceof RouteMatch) {
+        if (! ($matches instanceof RouteMatch || $matches instanceof V2RouteMatch)) {
             $this->fail('Failed to route!');
         }
 
@@ -140,7 +153,7 @@ class CollectionIntegrationTest extends TestCase
 
         $resource = new Resource();
         $events   = $resource->getEventManager();
-        $events->attach($this->listeners);
+        $this->listeners->attach($events);
 
         $controller = $this->controller = new RestController('Api\RestController');
         $controller->setResource($resource);
@@ -153,8 +166,11 @@ class CollectionIntegrationTest extends TestCase
 
     public function setUpContentNegotiation($controller)
     {
-        $plugins = new ControllerPluginManager();
-        $plugins->setService('Hal', $this->linksHelper);
+        $plugins = new ControllerPluginManager($this->prophesize(ContainerInterface::class)->reveal());
+        $plugins->setService('hal', $this->linksHelper);
+        if (method_exists($plugins, 'configure')) {
+            $plugins->setAlias('Hal', 'hal');
+        }
         $controller->setPluginManager($plugins);
 
         $viewModelSelector = $plugins->get('AcceptableViewModelSelector');
@@ -250,33 +266,13 @@ class CollectionIntegrationTest extends TestCase
         }
     }
 
-    public function createControllerManager(ServiceManager $container)
-    {
-        $r = new ReflectionMethod(ControllerManager::class, '__construct');
-        $args = $r->getParameters();
-        $first = array_shift($args);
-
-        if ($first->getName() === 'configuration') {
-            $controllers = new ControllerManager();
-            $controllers->setServiceLocator($container);
-            return $controllers;
-        }
-
-        return new ControllerManager($container);
-    }
-
     public function getServiceManager()
     {
         $services    = new ServiceManager();
-        $controllers = $this->createControllerManager($services);
-        $controllers->addAbstractFactory('ZF\Rest\Factory\RestControllerFactory');
-
-        $services->setService('Zend\ServiceManager\ServiceLocatorInterface', $services);
-        $services->setService('ControllerLoader', $controllers);
-        $services->setService('Config', [
+        $services->setService('config', [
             'zf-rest' => [
                 'Api\RestController' => [
-                    'listener'                   => 'CollectionIntegrationListener',
+                    'listener'                   => TestAsset\CollectionIntegrationListener::class,
                     'page_size'                  => 3,
                     'route_name'                 => 'resource',
                     'route_identifier_name'      => 'id',
@@ -285,36 +281,56 @@ class CollectionIntegrationTest extends TestCase
                 ],
             ],
         ]);
-        $services->setInvokableClass(
-            'SharedEventManager',
-            'Zend\EventManager\SharedEventManager'
-        );
-        $services->setInvokableClass(
-            'CollectionIntegrationListener',
-            'ZFTest\Rest\TestAsset\CollectionIntegrationListener'
-        );
-        $services->setFactory(
-            'EventManager',
-            'Zend\Mvc\Service\EventManagerFactory'
-        );
+
+        $services->setFactory('SharedEventManager', function ($container) {
+            return new SharedEventManager();
+        });
+
+        $services->setFactory(TestAsset\CollectionIntegrationListener::class, InvokableFactory::class);
+
+        $services->setFactory('EventManager', function ($container) {
+            $shared = $container->get('SharedEventManager');
+            if (method_exists($shared, 'getEvents')) {
+                $events = new EventManager();
+                $events->setSharedManager($shared);
+                return $events;
+            }
+
+            return new EventManager($shared);
+        });
+        $services->setShared('EventManager', false);
+
         $services->setFactory(
             'ControllerPluginManager',
             'Zend\Mvc\Service\ControllerPluginManagerFactory'
         );
-        $services->setShared('EventManager', false);
 
         $collection = $this->setUpCollection();
-        $services->addInitializer(function ($instance, $services) use ($collection) {
-            if (!$instance instanceof TestAsset\CollectionIntegrationListener) {
+        $services->addInitializer(function ($first, $second) use ($collection) {
+            // Initializer signature varies between v2 and v3
+            if ($first instanceof ServiceManager) {
+                // v3 signature
+                $instance = $second;
+            } else {
+                // v2 signature
+                $instance = $first;
+            }
+
+            if (! $instance instanceof TestAsset\CollectionIntegrationListener) {
                 return;
             }
             $instance->setCollection($collection);
         });
 
-        $controllers->setServiceLocator($services);
+        $controllers = new ControllerManager($services);
+        $controllers->addAbstractFactory('ZF\Rest\Factory\RestControllerFactory');
+        $services->setService('ControllerManager', $controllers);
 
         $plugins = $services->get('ControllerPluginManager');
-        $plugins->setService('Hal', $this->linksHelper);
+        $plugins->setService('hal', $this->linksHelper);
+        if (method_exists($plugins, 'configure')) {
+            $plugins->setAlias('Hal', 'hal');
+        }
 
         return $services;
     }
@@ -322,7 +338,7 @@ class CollectionIntegrationTest extends TestCase
     public function testFactoryEnabledListenerCreatesQueryStringWhitelist()
     {
         $services = $this->getServiceManager();
-        $controller = $services->get('ControllerLoader')->get('Api\RestController');
+        $controller = $services->get('ControllerManager')->get('Api\RestController');
         $controller->setEvent($this->getEvent());
         $this->setUpContentNegotiation($controller);
 
@@ -358,13 +374,13 @@ class CollectionIntegrationTest extends TestCase
     public function testFactoryEnabledListenerInjectsWhitelistedQueryParams()
     {
         $services = $this->getServiceManager();
-        $controller = $services->get('ControllerLoader')->get('Api\RestController');
+        $controller = $services->get('ControllerManager')->get('Api\RestController');
         $controller->setEvent($this->getEvent());
         $this->setUpContentNegotiation($controller);
 
         $controller->dispatch($this->request, $this->response);
         $resource = $controller->getResource();
-        $this->assertInstanceOf('ZF\Rest\Resource', $resource);
+        $this->assertInstanceOf(Resource::class, $resource);
         $params = $resource->getQueryParams();
 
         $this->assertInstanceOf('Zend\Stdlib\Parameters', $params);