+++ test/RouteMatchFactoryTrait.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
+ * @copyright Copyright (c) 2014 Zend Technologies USA Inc. (http://www.zend.com)
+ */
+
+namespace ZFTest\Rest;
+
+use Zend\Mvc\Router\RouteMatch as V2RouteMatch;
+use Zend\Router\RouteMatch;
+
+trait RouteMatchFactoryTrait
+{
+    /**
+     * Create and return a version-specific RouteMatch instance.
+     *
+     * @param array $params
+     * @return RouteMatch|V2RouteMatch
+     */
+    public function createRouteMatch(array $params = [])
+    {
+        $class = $this->getRouteMatchClass();
+        return new $class($params);
+    }
+
+    /**
+     * Return a version-specific route match class.
+     *
+     * @return string
+     */
+    public function getRouteMatchClass()
+    {
+        return class_exists(V2RouteMatch::class) ? V2RouteMatch::class : RouteMatch::class;
+    }
+}